Added early art and animations
I have changed the PlayerController script a tiny bit so that the player keeps facing the last direction they were in. Also to update animator.
I have added a walk anime and an idle animation to the player.
I have added the train (needs some tweaks), and backgrounds (still a lot of work to be done). I have also added animations for the background, train, and tried to add wind, but I feel like it is too distracting.